Test close_to within boundaries
