+++ tests/numeric_comparison_tests.py
@@ -1,6 +1,6 @@
 from nose.tools import istest, assert_equal
 
-from precisely import greater_than, greater_than_or_equal_to, less_than, less_than_or_equal_to, close_to
+from precisely import greater_than, greater_than_or_equal_to, less_than, less_than_or_equal_to
 from precisely.results import matched, unmatched
 
 
@@ -58,17 +58,3 @@ def less_than_or_equal_to_matches_when_actual_is_less_than_or_equal_to_value():
 def less_than_or_equal_to_description_describes_value():
     matcher = less_than_or_equal_to(42)
     assert_equal("less than or equal to 42", matcher.describe())
-
-
-@istest
-def close_to_matches_when_actual_is_close_to_value_plus_delta():
-    matcher = close_to(42, 1)
-    assert_equal(matched(), matcher.match(43))
-    assert_equal(matched(), matcher.match(41))
-    assert_equal(unmatched("was 40"), matcher.match(40))
-
-
-@istest
-def close_to_description_describes_value():
-    matcher = close_to(42, 1)
-    assert_equal("close to 42 +/- 1", matcher.describe())
